+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This changelog covers all three packages, as they are (for now) updated as a who
 - Make search result previews smaller
 - Fetch full collections when showing CollectionCard
 - `useResource` defaults to not accepting incomplete resources
+- Add `sign in` button to invite form
+- Improve UX in Tauri (desktop) mode
+  - Regular Links open in your browser, instead of in Atomic
 
 ## v0.30.0